Moved Speech sample to OpenVINO 2.0
